Add a CLI frontend to this jar, making this executatble
This supports a higher-level CLI tool for GoCD to allow syntax checking
on a pipeline config file. Currently, this supports 1 command, `syntax`
which will attempt to parse a specified file.

This also fixes the pipeline-export response deserialization issues on GoCD.
The `pipeline` value in the response needs to be double-serialized in order
for GoCD to deserialize it into a String.

This commit also switches the build tool for this plugin to use Gradle instead
of Maven, which lets us have better control in building a fat jar; we need to
explode all of the nested jars so that the classloader can load the nested classes
without any magic.